update flashtestation logic to use new workload ID computation
To account for a bug in the xfam and tdAttributes bitmasking logic, we changed the way the
workload ID is calculated in the BlockBuilderPolicy. That PR is here:

https://github.com/flashbots/flashtestations/pull/53

This means the workload ID calculation needs to change in any upstream services, which includes
op-rbuilder.

I've made the same change that I made in the solidity contract here in the attestation.rs logic
where op-rbuilder calculates the workload ID. The tests have been updated to use the new correct
workload ID, which I obtained by running the BlockBuilderPolicy:workloadIdForTDRegistration function
on the same TEE registration that is used in the tests.